dirty tricks
size: 4067
bench: 3573999

+++ src/chess.cpp
@@ -163,11 +163,11 @@ i16 move_make(i32 from, i32 to, i32 promo = PAWN) {
     return from | to << 6 | promo << 12;
 }
 
-i32 move_from(i16 move) {
+i32 move_from(i32 move) {
     return move & 63;
 }
 
-i32 move_to(i16 move) {
+i32 move_to(i32 move) {
     return move >> 6 & 63;
 }
